+++ src/gmoor/LineSegment.h
@@ -5,9 +5,9 @@ class LineSegment
 	double m_length;
 	size_t m_property;  //Property ID
 	unsigned int m_discretization;
-	unsigned int m_tot_nodes;
-	unsigned int m_tot_elements;
-	unsigned int m_nodeset;  //TODO:  definir o nodeset do segmento
+	size_t m_tot_nodes;
+	size_t m_tot_elements;
+	size_t m_nodeset;  //TODO:  definir o nodeset do segmento
  	double m_epsilon;  // Epsilon used to establish contact
 	double m_gamma;  // immersed  
 
@@ -25,9 +25,9 @@ public:
 	void SetLength(double length);
 	void SetProperty(size_t property);
 	void SetDiscretization(unsigned int discretization);
-	void SetNNodes(unsigned int nodes);
-	void SetNElements(unsigned int elements);
-	void SetNodeSet(unsigned int nodeset);
+	void SetNNodes(size_t nodes);
+	void SetNElements(size_t elements);
+	void SetNodeSet(size_t nodeset);
 	void SetEpsilon(double epsilon);
 	void SetGamma(double gamma);
 
@@ -39,9 +39,9 @@ public:
 	inline double GetLength() const					{ return this->m_length; }
 	inline size_t GetProperty() const				{ return this->m_property; }
 	inline unsigned int GetDiscretization() const	{ return this->m_discretization; }
-	inline unsigned int GetNNodes() const			{ return this->m_tot_nodes; }
-	inline unsigned int GetNElements() const		{ return this->m_tot_elements; }
-	inline unsigned int GetNodeSet() const			{ return this->m_nodeset; }
+	inline size_t GetNNodes() const					{ return this->m_tot_nodes; }
+	inline size_t GetNElements() const				{ return this->m_tot_elements; }
+	inline size_t GetNodeSet() const				{ return this->m_nodeset; }
 	inline double GetEpsilon() const				{ return this->m_epsilon; }
 	inline double GetGamma() const					{ return this->m_gamma; }
