+++ src/gmoor/Vessel.h
@@ -24,11 +24,11 @@ private:
 	unsigned int m_element;
 
 	// Concentrated mass
-	float m_mass;
+	double m_mass;
 	
 	// Inertia tensor
 	/// [ J11, J22, J33, J12, J13, J23 ]
-	std::array<float, 6> m_inertia_tensor;
+	std::array<double, 6> m_inertia_tensor;
 
 	// Contains the nodes to couple vessel and lines 
 	std::unordered_set<unsigned int> m_nodes_rigid_nodeset;
@@ -53,8 +53,8 @@ public:
 	void SetNodeset(unsigned int nodeset);
 	void SetRigidNodeset(unsigned int rigid_nodeset);
 	void SetElement(unsigned int element);
-	void SetMass(float mass);
-	void SetInertiaTensor(const std::array<float, 6>& inertia_tensor);
+	void SetMass(double mass);
+	void SetInertiaTensor(const std::array<double, 6>& inertia_tensor);
 	void SetNodesRigidNodeset(const std::unordered_set<unsigned int>& nodes_rigid_nodeset);
 
 
@@ -68,9 +68,9 @@ public:
 	inline unsigned int GetNodeset() const		{ return this->m_nodeset; }
 	inline unsigned int GetRigidNodeset() const	{ return this->m_rigid_nodeset; }
 	inline unsigned int GetElement() const		{ return this->m_element; }
-	inline float GetMass() const				{ return this->m_mass; }
+	inline double GetMass() const				{ return this->m_mass; }
 
-	inline std::array<float, 6> const& GetInertiaTensor() const					{ return this->m_inertia_tensor; }
+	inline std::array<double, 6> const& GetInertiaTensor() const					{ return this->m_inertia_tensor; }
 	inline std::unordered_set<unsigned int> const& GetNodesRigidNodeset() const { return this->m_nodes_rigid_nodeset; }
 	inline std::unordered_set<unsigned int>& GetNodesRigidNodeset()				{ return this->m_nodes_rigid_nodeset; }
